improved sequence file logic (#95)
