+++ src/main/steps/applicant2/applicant2Sequence.ts
@@ -1,18 +1,6 @@
-import { CaseWithId } from '../../app/case/case';
 import { YesOrNo } from '../../app/case/definition';
+import { Sections, Step } from '../constants';
 import * as Urls from '../urls';
-
-export enum Sections {
-  AboutApplicant2 = 'aboutApplicant2',
-}
-
-export interface Step {
-  url: string;
-  showInSection?: Sections;
-  showInCompleteSection?: Sections;
-  excludeFromContinueApplication?: boolean;
-  getNextStep: (data: Partial<CaseWithId>) => Urls.PageLink;
-}
 
 export const applicant2Sequence: Step[] = [
   {
