+++ src/main/steps/review-pay-submit/reviewPaySubmitSequence.ts
@@ -1,17 +1,5 @@
-import { CaseWithId } from '../../app/case/case';
+import { Sections, Step } from '../constants';
 import * as Urls from '../urls';
-
-export enum Sections {
-  ReviewPaySubmit = 'reviewPaySubmit',
-}
-
-export interface Step {
-  url: string;
-  showInSection?: Sections;
-  showInCompleteSection?: Sections;
-  excludeFromContinueApplication?: boolean;
-  getNextStep: (data: Partial<CaseWithId>) => Urls.PageLink;
-}
 
 export const reviewPaySubmitSequence: Step[] = [
   {
